Adding ideas into the database without user constraints

+++ src/main/java/com/idea/share/com/idea/share/user/User.java
@@ -1,5 +1,6 @@
 package com.idea.share.com.idea.share.user;
 
+import com.idea.share.com.idea.share.idea.Idea;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -7,7 +8,10 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
+import java.util.ArrayList;
+import java.util.List;
 
 @Data
 @AllArgsConstructor
@@ -17,9 +21,11 @@ import javax.persistence.Table;
 @Table(name = "users")
 public class User {
     @Id
-    private Long id;
+    private Integer id;
     private String name;
     private String email;
     private String password;
+//    @OneToMany(mappedBy = "user")
+//    private List<Idea> ideas = new ArrayList<>();
 
 }
